+++ game/target.cpp
@@ -861,7 +861,7 @@ idTarget_SetInfluence::Event_Flash
 */
 void idTarget_SetInfluence::Event_Flash( float flash, int out ) {
 	idPlayer *player = gameLocal.GetLocalPlayer();
-	player->playerView.Fade( idVec4( 1, 1, 1, 1 ), flash );
+	player->playerView.Fade( idVec4( 1, 1, 1, 1 ), static_cast<int>(flash) );
 	const idSoundShader *shader = NULL;
 	if ( !out && flashInSound.Length() ){
 		shader = declManager->FindSound( flashInSound );
@@ -881,7 +881,7 @@ idTarget_SetInfluence::Event_ClearFlash
 */
 void idTarget_SetInfluence::Event_ClearFlash( float flash ) {
 	idPlayer *player = gameLocal.GetLocalPlayer();
-	player->playerView.Fade( vec4_zero , flash );		
+	player->playerView.Fade( vec4_zero , static_cast<int>(flash) );
 }
 /*
 ================
@@ -1319,7 +1319,7 @@ void idTarget_SetFov::Event_Activate( idEntity *activator ) {
 	cinematic = true;
 
 	idPlayer *player = gameLocal.GetLocalPlayer();
-	fovSetting.Init( gameLocal.time, SEC2MS( spawnArgs.GetFloat( "time" ) ), player ? player->DefaultFov() : g_fov.GetFloat(), spawnArgs.GetFloat( "fov" ) );
+	fovSetting.Init( gameLocal.time, SEC2MS( spawnArgs.GetFloat( "time" ) ), static_cast<int>(player ? player->DefaultFov() : g_fov.GetFloat()), static_cast<int>(spawnArgs.GetFloat( "fov" )) );
 	BecomeActive( TH_THINK );
 }
